+++ src/wasm32_utils.rs
@@ -1,5 +1,4 @@
 use std::arch::wasm32::*;
-use std::ptr;
 
 use crate::pixels::{U8x3, U8x4};
 
@@ -12,45 +11,43 @@ pub(crate) unsafe fn load_v128<T>(buf: &[T], index: usize) -> v128 {
 #[inline]
 #[target_feature(enable = "simd128")]
 pub(crate) unsafe fn loadl_i64<T>(buf: &[T], index: usize) -> v128 {
-    let i = buf.get_unchecked(index..).as_ptr() as *const i64;
-    i64x2(ptr::read_unaligned(i), 0)
+    let p = buf.get_unchecked(index..).as_ptr() as *const i64;
+    i64x2(p.read_unaligned(), 0)
 }
 
 #[inline]
 #[target_feature(enable = "simd128")]
 pub(crate) unsafe fn loadl_i32<T>(buf: &[T], index: usize) -> v128 {
-    let i = buf.get_unchecked(index..).as_ptr() as *const i32;
-    i32x4(ptr::read_unaligned(i), 0, 0, 0)
+    let p = buf.get_unchecked(index..).as_ptr() as *const i32;
+    i32x4(p.read_unaligned(), 0, 0, 0)
 }
 
 #[inline]
 #[target_feature(enable = "simd128")]
 pub(crate) unsafe fn loadl_i16<T>(buf: &[T], index: usize) -> v128 {
-    let i = buf.get_unchecked(index..).as_ptr() as *const i16;
-    i16x8(ptr::read_unaligned(i), 0, 0, 0, 0, 0, 0, 0)
+    let p = buf.get_unchecked(index..).as_ptr() as *const i16;
+    i16x8(p.read_unaligned(), 0, 0, 0, 0, 0, 0, 0)
 }
 
 #[inline]
 #[target_feature(enable = "simd128")]
 pub(crate) unsafe fn ptr_i16_to_set1_i64(buf: &[i16], index: usize) -> v128 {
-    i64x2_splat(ptr::read_unaligned(
-        buf.get_unchecked(index..).as_ptr() as *const i64
-    ))
+    let p = buf.get_unchecked(index..).as_ptr() as *const i64;
+    i64x2_splat(p.read_unaligned())
 }
 
 #[inline]
 #[target_feature(enable = "simd128")]
 pub(crate) unsafe fn ptr_i16_to_set1_i32(buf: &[i16], index: usize) -> v128 {
-    i32x4_splat(ptr::read_unaligned(
-        buf.get_unchecked(index..).as_ptr() as *const i32
-    ))
+    let p = buf.get_unchecked(index..).as_ptr() as *const i32;
+    i32x4_splat(p.read_unaligned())
 }
 
 #[inline]
 #[target_feature(enable = "simd128")]
 pub(crate) unsafe fn i32x4_extend_low_ptr_u8(buf: &[u8], index: usize) -> v128 {
-    let ptr = buf.get_unchecked(index..).as_ptr() as *const v128;
-    u32x4_extend_low_u16x8(i16x8_extend_low_u8x16(v128_load(ptr)))
+    let p = buf.get_unchecked(index..).as_ptr() as *const v128;
+    u32x4_extend_low_u16x8(i16x8_extend_low_u8x16(v128_load(p)))
 }
 
 #[inline]
@@ -70,8 +67,8 @@ pub(crate) unsafe fn i32x4_extend_low_ptr_u8x3(buf: &[U8x3], index: usize) -> v1
 #[inline]
 #[target_feature(enable = "simd128")]
 pub(crate) unsafe fn i32x4_v128_from_u8(buf: &[u8], index: usize) -> v128 {
-    let ptr = buf.get_unchecked(index..).as_ptr() as *const i32;
-    i32x4(*ptr, 0, 0, 0)
+    let p = buf.get_unchecked(index..).as_ptr() as *const i32;
+    i32x4(p.read_unaligned(), 0, 0, 0)
 }
 
 #[inline]