+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Crate
 
-- Added using of (read|write)_unaligned for unaligned pointers on `arm64` architecture.
+- Added using of (read|write)_unaligned for unaligned pointers 
+  on `arm64` and `wasm32` architecture.
   ([#15](https://github.com/Cykooz/fast_image_resize/issues/15)).
 
 ## [2.7.1] - 2023-04-28